+++ src/Command/ScaffoldThemeCommand.php
@@ -2,9 +2,11 @@
 
 use Composer\Command\BaseCommand;
 use Composer\Composer;
+use Composer\Json\JsonFile;
 use Composer\Package\Package;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
@@ -26,7 +28,9 @@ class ScaffoldThemeCommand extends BaseCommand {
 	protected function configure() {
 		$this->setName( 'scaffold-theme' )
 		     ->setDescription( 'Bootstrap a new WordPress theme using Be API\'s frontend framework.' )
-		     ->addArgument( 'folder', InputArgument::REQUIRED, "Your theme's folder name" );
+		     ->addArgument( 'folder', InputArgument::REQUIRED, "Your theme's folder name" )
+		     ->addOption( 'boilerplate-version', null, InputOption::VALUE_OPTIONAL, 'Wich version to use ?', 'Latest' )
+		     ->addOption( 'no-autoload', null, InputOption::VALUE_NONE, 'Autoload the class into composer.json' );
 	}
 
 	protected function execute( InputInterface $input, OutputInterface $output ) {
@@ -34,6 +38,8 @@ class ScaffoldThemeCommand extends BaseCommand {
 		$io        = new SymfonyStyle( $input, $output );
 		$composer  = $this->getComposer();
 		$themeName = $input->getArgument( 'folder' );
+		$version    = $input->getOption( 'boilerplate-version' );
+		$no_autoload = $input->getOption( 'no-autoload' );
 
 		// what is the command's purpose
 		$io->write( "\nHello, this command allows you to start a theme with the wonderful Be API theme Framework." );
@@ -66,8 +72,10 @@ class ScaffoldThemeCommand extends BaseCommand {
 			$io->write( "You did not provide any real name so I take the folder name by default." );
 			$themeCompleteName = $themeName;
 		}
-		$this->generateTheme( $composer, $io, $themeName, $themePath, $themeCompleteName, $downloadPath );
-		$io->write( "\nYour theme is ready ! :)" );
+
+		$this->generateTheme( $composer, $io, $themeName, $themePath, $themeCompleteName, $downloadPath, $output, $version, $no_autoload );
+		$io->success( "\nYour theme is ready ! :)" );
+		$io->success( 'Run composer dump-autoload to make the autoloading work :)' );
 	}
 
 	/**
@@ -172,12 +180,17 @@ class ScaffoldThemeCommand extends BaseCommand {
 	 * @param $themeCompleteName
 	 * @param $downloadPath
 	 *
+	 * @param $output
+	 * @param $version
+	 * @param $no_autoload
+	 *
+	 * @throws \Exception
 	 * @author Julien Maury
 	 */
-	protected function generateTheme( $composer, $io, $themeName, $themePath, $themeCompleteName, $downloadPath ) {
+	protected function generateTheme( $composer, $io, $themeName, $themePath, $themeCompleteName, $downloadPath, $output, $version, $no_autoload ) {
 
 		if ( ! file_exists( $downloadPath . '/index.php' ) ) {
-			$composer->getDownloadManager()->download( $this->getThemePackage(), $downloadPath );
+			$composer->getDownloadManager()->download( $this->getThemePackage( $version ), $downloadPath );
 		}
 
 		if ( ! file_exists( $downloadPath . '/index.php' ) ) {
@@ -193,18 +206,50 @@ class ScaffoldThemeCommand extends BaseCommand {
 
 		$this->doStrReplace( $themePath, 'BEA\\Theme\\Framework', $themeNamespace );
 		$this->replaceHeaderStyle( $themePath, static::$search, $themeCompleteName );
+
+		/**
+		 * Add the new namespace to the autoload entry of the composer.json file.
+		 *
+		 */
+		if ( false === $no_autoload ) {
+			$composerPath = $composer->getConfig()->getConfigSource()->getName();
+			$composerFile = new JsonFile( $composerPath );
+
+			try {
+				$composerJson = $composerFile->read();
+				$composerJson['autoload']['psr-4'][$themeNamespace."\\"] = $themePath.'/inc/';
+
+
+				$composerFile->write( $composerJson );
+				$output->writeln( "The namespace have been added to the composer.json file !" );
+			} catch ( RuntimeException $e ) {
+				$output->writeln( "<error>An error occurred</error>" );
+				$output->writeln( sprintf( "<error>%s</error>", $e->getMessage() ) );
+				exit;
+			}
+		}
+
 	}
 
 	/**
 	 * Setup a dummy package for Composer to download
 	 *
+	 * @param $version
+	 *
 	 * @return Package
 	 */
-	protected function getThemePackage() {
-		$p = new Package( 'theme-framework', 'starter-php', 'Latest' );
+	protected function getThemePackage( $version ) {
+		$p = new Package( 'theme-framework', 'starter-php', $version );
 		$p->setType( 'library' );
 		$p->setDistType( 'zip' );
-		$p->setDistUrl( self::$zip_url );
+
+		$dist_url = self::$zip_url;
+
+		if (  'Latest' !== $version ) {
+			$dist_url = sprintf( 'https://github.com/BeAPI/beapi-frontend-framework/archive/%s.zip', $version );
+		}
+
+		$p->setDistUrl( $dist_url );
 
 		return $p;
 	}