+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog ##
 
+## 1.2.0 - 5 May 2020
+* Handle psr-4 and no psr-4 themes
+* Add the autoload of files directly to the composer.json file
+* Add the --no-autoload option
+* Add the --boilerplate-version option
+
 ## 1.1.0 - 17 Dec 2019
 * Fix hardcoded path when scaffolding a new theme